Move WizardStep into StepsAccordion and move StepsAccordion into layout
